Cope with blank metadata values better. Fixes GH-3.
The Java version has a pair of fields (e.g. hasNationalNumberPattern 
and nationalNumberPattern_) to indicate presence of the field 
separately from field value, and the field value defaults to ""
even when it is not present.

The Python version uses a single field that takes the value None
to indicate field not present.  However, the rest of the code does
sometimes use the field even when it is not present, relying on
the fallback default value "".