buildmetadata: relax consistency checks with --lax
In various places the metadata does not comply with the schema
associated with the XML, so relax the checks for this.

Examples:
 - Some number descriptions have no required possibleLengths
   element (example: GE.noInternationalDialling).
 - Some number descriptions have no required exampleNumber
   element (example: MC.noInternationalDialling).
 - Some number descriptions have no required nationalNumberPattern
   element (example: testdata.BB.fixedLine).
 - Some generalDesc number descriptions have an exampleNumber element
   which isn't in the schema (example: testdata.AM.generalDesc).

Control this with --lax, with the aim that the --lax flag can be
removed for the real metadat in future.  (It will probably still
be required for the test metadata.)