Add imports of int values to .pyi files
A statement like `from .local import THING` in `current.py` makes
`THING` available in the current module, but this also acts as ani
export (i.e. `current.THING` is defined).

As of version 0.930 of mypy, the `stubtest` checker wants to see a
type definition for such things, when they are ints or strings.

Add the equivalent `from .local import THING` statements into the
.pyi stub files, so `stubtest` can see the associated type definition.

For autogenerated `*data/__init__.pyi` files, this involves a change to
the metadata generation tool.

Also:
 - Fix asyoutypeformatter.py to import `REGION_CODE_FOR_NON_GEO_ENTITY`
   from `phonemetadata` (the definition) rather than `phonenumberutil`
   (a re-export).
 - Rename local loop variables to start with an underscore in
   (autogenerator for) `*data/__init__.py` so they don't look like
   things that should be exported.
 - Drop version pin (==0.921) for mypy.  Fixes #224.